Restructuring project with better separation of modules
