+++ src/core/naive.test.js
@@ -1,0 +1,7 @@
+import { describe, it, expect } from 'vitest'
+
+describe.todo('function #arePixelsEqual')
+describe.todo('function #highlightedDiffPixel')
+describe.todo('function #diffResultMessage')
+describe.todo('function #calcDiffWithCanvas')
+describe.todo('function #testImagesDiff')
